changed deprecated by, description and moved TTLTest from apoc.date to apoc.ttl
